% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_magicf.R
\name{assess_magicf}
\alias{assess_magicf}
\title{Computes statistics for the magic formula}
\usage{
assess_magicf(tikr, mrkt)
}
\arguments{
\item{tikr}{tiker of the company, char.}

\item{mrkt}{name of the market, char.}
}
\description{
Computes statistics for the magic formula
}
