% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_ruleone.R
\name{assess_ruleone}
\alias{assess_ruleone}
\title{Computes the rule-one statistics}
\usage{
assess_ruleone(tikr, mrkt, nyrs = 5)
}
\arguments{
\item{tikr}{tiker of the company, char.}

\item{mrkt}{the name of the market, char.}

\item{nyrs}{number of years to analyze, int. (Default, 5)}
}
\description{
Computes the rule-one statistics
}
