% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_sheet_statements.R
\name{balance_sheet_statement}
\alias{balance_sheet_statement}
\title{Get balance sheet statements}
\usage{
balance_sheet_statement(ticker, period = "annualy", api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{period}{Required. The time period.
Valid values are:
"annualy" Default
"quarterly"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing company balance sheet statements data.
}
\description{
Get balance sheet statements
}
\examples{
balance_sheet_statement(ticker = "AAPL", period = "annualy", api_key = api)
}
