% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash_flow_statement.R
\name{cash_flow_statement}
\alias{cash_flow_statement}
\title{Get company cash flow statement}
\usage{
cash_flow_statement(ticker, period = "annualy", api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{period}{Required. The time period.
Valid values are:
"annualy" Default
"quarterly"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing company cash flow statement data.
}
\description{
Get company cash flow statement
}
\examples{
cash_flow_statement(
ticker = "AAPL",
period = "annualy",
api_key = "42632c388365783ab0c341411f47ae80")
}
