% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cik_list.R
\name{cik_list}
\alias{cik_list}
\title{Get cik list}
\usage{
cik_list(api_key = api)
}
\arguments{
\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing cik list.
}
\description{
Get cik list
}
\examples{
cik_list(api_key = "42632c388365783ab0c341411f47ae80")
}
