% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_enterprise_value.R
\name{company_enterprise_value}
\alias{company_enterprise_value}
\title{Get company enterprise value}
\usage{
company_enterprise_value(ticker, period = "annualy", api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{period}{Required. The time period.
Valid values are:
"annualy" Default
"quarterly"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing company enterprise value data.
}
\description{
Get company enterprise value
}
\examples{
company_enterprise_value(ticker = "AAPL", period = "annualy", api_key = api)
}
