% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_financial_growth.R
\name{company_financial_growth}
\alias{company_financial_growth}
\title{Get company financial growth}
\usage{
company_financial_growth(ticker, period = "annualy", api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{period}{Required. The time period.
Valid values are:
"annualy" Default
"quarterly"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing company financial growth data.
}
\description{
Get company financial growth
}
\examples{
company_financial_growth(ticker = "AAPL", period = "annualy", api_key = api)
}
