% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_financial_ratios.R
\name{company_financial_ratios}
\alias{company_financial_ratios}
\title{Get company financial ratios}
\usage{
company_financial_ratios(tickers, api_key = api)
}
\arguments{
\item{tickers}{Required. The name of the tickers.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing company financial ratios data.
}
\description{
Get company financial ratios
}
\examples{
company_financial_ratios(tickers = c("AAPL"), api_key = "42632c388365783ab0c341411f47ae80")
}
