% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_financial_statements.R
\name{company_financial_statements}
\alias{company_financial_statements}
\title{Get company financial statements}
\usage{
company_financial_statements(ticker, period = "annualy", api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{period}{Required. The time period.
Valid values are:
"annualy" Default
"quarterly"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing company financial statements data.
}
\description{
Get company financial statements
}
\examples{
company_financial_statements(ticker = "AAPL", period = "annualy")
}
