% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_key_metrics.R
\name{company_key_metrics}
\alias{company_key_metrics}
\title{Get company key metrics}
\usage{
company_key_metrics(tickers, api_key = api)
}
\arguments{
\item{tickers}{Required. The name of the tickers.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing company key metrics data.
}
\description{
Get company key metrics
}
\examples{
company_key_metrics(tickers = c("AAPL"), api_key = "42632c388365783ab0c341411f47ae80")
}
