% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_ratings.R
\name{company_ratings}
\alias{company_ratings}
\title{Get company ratings}
\usage{
company_ratings(tickers, api_key = api)
}
\arguments{
\item{tickers}{Required. The name of the tickers.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing company ratings data.
}
\description{
Get company ratings
}
\examples{
company_ratings(tickers = c("AAPL", "MSFT"), api_key = api)
}
