% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_indicators.R
\name{daily_indicators}
\alias{daily_indicators}
\title{Get daily indicators}
\usage{
daily_indicators(ticker, period, type, api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{period}{Required. Period for daily indicators}

\item{type}{Rquired. Type of daily indicators.
Valid values are:
"SMA"
"EMA"
"WMA"
"DEMA"
"TEMA"
"williams"
"RSI"
"ADX"
"standardDeviation"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing daily indicators.
}
\description{
Get daily indicators
}
\examples{
daily_indicators(ticker = "AAPL",
period = 10,
type = "EMA",
api_key = "42632c388365783ab0c341411f47ae80")
}
