% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounted_cash_flow.R
\name{discounted_cash_flow}
\alias{discounted_cash_flow}
\title{Get discounted cash flow data}
\usage{
discounted_cash_flow(tickers, api_key = api)
}
\arguments{
\item{tickers}{Required. The name of the tickers.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing discounted cash flow data.
}
\description{
Get discounted cash flow data
}
\examples{
discounted_cash_flow(tickers = c("AAPL"), api_key = "42632c388365783ab0c341411f47ae80")
}
