% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earning_calendar.R
\name{earning_calendar}
\alias{earning_calendar}
\title{Get earning calendar}
\usage{
earning_calendar(ticker, datefrom, dateto, api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{datefrom}{Required. Start date in format "yyyy-mm-dd"}

\item{dateto}{Rquired. End date in format "yyyy-mm-dd"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing earning calendar.
}
\description{
Get earning calendar
}
\examples{
earning_calendar(ticker = "AAPL",
datefrom = "2019-01-01",
dateto = "2020-01-01",
api_key = "42632c388365783ab0c341411f47ae80")
}
