% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etf_holders.R
\name{etf_holders}
\alias{etf_holders}
\title{Get etf holders}
\usage{
etf_holders(ticker, api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing etf holders.
}
\description{
Get etf holders
}
\examples{
etf_holders(ticker = "AAPL", api_key = api)
}
