% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financeR.R
\docType{package}
\name{financeR-package}
\alias{financeR}
\alias{financeR-package}
\title{Finance with R - Tutorial Codes}
\description{
Open any tutorial code from "Finance with R". Furthermore update the tutorial contents anytime without updating the package.
}
\references{
\url{http://www.finance-r.com/}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.finance-r.com/}
}

}
\author{
Ronald Hochreiter, \email{ron@hochreiter.net}
}
