% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_symbol_list.R
\name{financial_statement_symbol_list}
\alias{financial_statement_symbol_list}
\title{Get finanial statement symbol list}
\usage{
financial_statement_symbol_list(api_key = api)
}
\arguments{
\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing financial statement symbol list
}
\description{
Get finanial statement symbol list
}
\examples{
financial_statement_symbol_list(api_key = "42632c388365783ab0c341411f47ae80")
}
