% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_statements_growth.R
\name{financial_statements_growth}
\alias{financial_statements_growth}
\title{Get financial statement growth data}
\usage{
financial_statements_growth(ticker, type, api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{type}{Rquired. Type of financial statement growth
Valid values are:
"income"
"balance_sheet"
"cash_flow"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing company financial statement growth data.
}
\description{
Get financial statement growth data
}
\examples{
financial_statements_growth(
ticker = "AAPL",
type = "income",
api_key = "42632c388365783ab0c341411f47ae80")
}
