% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_balance_sheet.R
\name{get_balance_sheet}
\alias{get_balance_sheet}
\title{This function imports the balance sheet of the equity from the 10-Q.}
\usage{
get_balance_sheet(stock, fy, form = "10-K", fp = "FY")
}
\arguments{
\item{stock}{stock ticker to lookup such as ML for NYSE:MONEYLION}

\item{fy}{fiscal year for the period to look up such as 2021}

\item{form}{the form to scrape  from edgar such as 10-Q or 10-K}

\item{fp}{fiscal period within the fiscal year such as Q1 or Q3 or at times FY if its for the whole year}
}
\value{
a dataframe with the balance sheet
}
\description{
This function imports the balance sheet of the equity from the 10-Q.
}
