% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_income_statement.R
\name{get_income_statement}
\alias{get_income_statement}
\title{This function imports the income statement of the equity from the 10-Q.}
\usage{
get_income_statement(stock, fy, form = "10-K", fp = "FY")
}
\arguments{
\item{stock}{stock ticker to lookup such as ML for NYSE:MONEYLION}

\item{fy}{fiscal year for the period to look up such as 2021}

\item{form}{the form to scrape  from edgar such as 10-Q or 10-K}

\item{fp}{fiscal period within the fiscal year such as Q1 or Q3 or at times FY if its for the whole year}
}
\value{
a dataframe with the income statement
}
\description{
This function imports the income statement of the equity from the 10-Q.
}
