% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cashflow_statement.R,
%   R/get_stockholder_equity_statement.R
\name{get_statement_of_cashflows}
\alias{get_statement_of_cashflows}
\title{This function imports the cashflow statement of the equity from the 10-Q.}
\usage{
get_statement_of_cashflows(stock, fy, form = "10-K", fp = "FY")

get_statement_of_cashflows(stock, fy, form = "10-K", fp = "FY")
}
\arguments{
\item{stock}{stock ticker to lookup such as ML for NYSE:MONEYLION}

\item{fy}{fiscal year for the period to look up such as 2021}

\item{form}{the form to scrape  from edgar such as 10-Q or 10-K}

\item{fp}{fiscal period within the fiscal year such as Q1 or Q3 or at times FY if its for the whole year}
}
\value{
a dataframe with the statement of cashflows

a dataframe with the statement of cashflows
}
\description{
This function imports the cashflow statement of the equity from the 10-Q.

This function imports the stockholder equity statement of the equity from the 10-Q.
}
