% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tickers_vector.R
\name{get_tickers_vector}
\alias{get_tickers_vector}
\title{Get tickers as vector}
\usage{
get_tickers_vector(api_key = api)
}
\arguments{
\item{api_key}{Required. Your API key.}
}
\value{
Returns a vector of results containing all available tickers
}
\description{
Get tickers as vector
}
