% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic_price.R
\name{historic_price}
\alias{historic_price}
\title{Get historic stock price}
\usage{
historic_price(ticker, time, api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{time}{Required. The time of which the market is updated. full, 1min, 5min, 15min, 30min, 1hour or 4hour.
Valid values are:
"full"
"1min"
"5min"
"15min"
"30min"
"1hour"
"4hour"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing historic stock price data.
}
\description{
Get historic stock price
}
\examples{
historic_price(ticker = "AAPL", time = "full", api_key = "42632c388365783ab0c341411f47ae80")

}
