% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_dcf.R
\name{historical_dcf}
\alias{historical_dcf}
\title{Get historical discounted cash flow data}
\usage{
historical_dcf(tickers, api_key = api)
}
\arguments{
\item{tickers}{Required. The name of the tickers.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing historical discounted cash flow data.
}
\description{
Get historical discounted cash flow data
}
\examples{
historical_dcf(tickers = c("AAPL", "MSFT"), api_key = api)
}
