% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_dividend.R
\name{historical_dividends}
\alias{historical_dividends}
\title{Get historical dividend}
\usage{
historical_dividends(ticker, api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing historical dividend data.
}
\description{
Get historical dividend
}
\examples{
historical_dividends(ticker = "AAPL", api_key = api)
}
