% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instituational_holders.R
\name{institutional_holders}
\alias{institutional_holders}
\title{Get institutional holders}
\usage{
institutional_holders(ticker, api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing institutional holders.
}
\description{
Get institutional holders
}
\examples{
institutional_holders(ticker = "AAPL", api_key = "42632c388365783ab0c341411f47ae80")
}
