% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday_indicators.R
\name{intraday_indicators}
\alias{intraday_indicators}
\title{Get in trade day indicators}
\usage{
intraday_indicators(ticker, time, period, type, api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{time}{Rquired. The time-frequency.
Valid values are:
"1min"
"5min"
"15min"
"30min"
"1hour"
"4hour"}

\item{period}{Required. Period for daily indicators}

\item{type}{Rquired. Type of daily indicators.
Valid values are:
"SMA"
"EMA"
"WMA"
"DEMA"
"TEMA"
"williams"
"RSI"
"ADX"
"standardDeviation"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing in trade day indicators.
}
\description{
Get in trade day indicators
}
\examples{
intraday_indicators(ticker = "AAPL",
time = "1hour",
period =  10,
type = "SMA",
api_key = "42632c388365783ab0c341411f47ae80")
}
