% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_executives.R
\name{key_executives}
\alias{key_executives}
\title{Get company key executives}
\usage{
key_executives(tickers, api_key = api)
}
\arguments{
\item{tickers}{Required. The name of the tickers.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing company key executives data.
}
\description{
Get company key executives
}
\examples{
key_executives(tickers = "AAPL", api_key = "42632c388365783ab0c341411f47ae80")
}
