% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/major_index_list.R
\name{major_index_list}
\alias{major_index_list}
\title{Get major index list}
\usage{
major_index_list(api_key = api)
}
\arguments{
\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing major index list data.
}
\description{
Get major index list
}
\examples{
major_index_list(api_key = "42632c388365783ab0c341411f47ae80")
}
