% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_capitalization.R
\name{market_capitalization}
\alias{market_capitalization}
\title{Get market capitalixation}
\usage{
market_capitalization(tickers, api_key = api)
}
\arguments{
\item{tickers}{Required. The name of the tickers.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing market capitalization data.
}
\description{
Get market capitalixation
}
\examples{
market_capitalization(tickers = c("AAPL", "MSFT"), api_key = api)
}
