% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morning_keyratios.R
\name{morning_keyrations}
\alias{morning_keyrations}
\title{Returns the key ratios of the company}
\usage{
morning_keyrations(tikr, mrkt)
}
\arguments{
\item{mrkt}{}
}
\description{
Returns the key ratios of the company
}
