% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morning_report.R
\name{morning_report}
\alias{morning_report}
\title{Gets financial reports from Morningstar}
\usage{
morning_report(tikr, mrkt, rprt)
}
\arguments{
\item{tikr}{tiker of the company, char.}

\item{mrkt}{name of the market, char.}

\item{rprt}{name of the report, char.}
}
\description{
For the market (\code{mrkt}), you can choose between
\code{"XNAS"} (Nasdaq Exchange),
\code{"XASE"} (American Stock Exchange),
\code{"XNYS"} (New York's Stock Exchange),
\code{"XHKG"} (Hong Kong Stock Exchange),
\code{"XSHE"} (ShenZhen Stock Exchange), and
\code{"XSHG"} (Shangai Stock Exchange).
For the report, please select among
\code{"is"} (income statement),
\code{"cf"} (cash flow), and
\code{"bs"} (balance sheet).
}
\examples{

morning_report("ILMN", "XNAS", "bs")

}
