% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_active.R
\name{most_active}
\alias{most_active}
\title{Get most active stocks}
\usage{
most_active(api_key = api)
}
\arguments{
\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing most active stocks.
}
\description{
Get most active stocks
}
\examples{
most_active(api_key = "42632c388365783ab0c341411f47ae80")
}
