% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_loser.R
\name{most_loser}
\alias{most_loser}
\title{Get most losing stocks}
\usage{
most_loser(api_key = api)
}
\arguments{
\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing most losing stocks.
}
\description{
Get most losing stocks
}
\examples{
most_loser(api_key = "42632c388365783ab0c341411f47ae80")
}
