% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninentynine_stocklist.R
\name{ninentynine_stocklist}
\alias{ninentynine_stocklist}
\title{Downloads the companies traded by 99}
\usage{
ninentynine_stocklist(retrn = TRUE)
}
\description{
Downloads the companies traded by 99
}
