% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickfs_findmrkt.R
\name{quickfs_findmrkt}
\alias{quickfs_findmrkt}
\title{Finds the country and market of a tikr}
\usage{
quickfs_findmrkt(tikr, tikrs)
}
\arguments{
\item{tikr}{tiker of the company, char.}

\item{tikrs}{a list of tikrs, from \code{\link{quickfs_getikrs()}}}
}
\description{
Finds the country and market of a tikr
}
\examples{

tikrs = quickfs_getikrs()
quickfs_findmrkt("ILMN", tikrs)

}
