% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickfs_getikrs.R
\name{quickfs_getikrs}
\alias{quickfs_getikrs}
\title{Gets the list of companies in different markets}
\usage{
quickfs_getikrs()
}
\description{
Gets the list of companies in different markets
}
\examples{

tikrs = quickfs_getikrs()

}
