% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickfs_getmetrics.R
\name{quickfs_getmetric}
\alias{quickfs_getmetric}
\title{Retrieve financial metrics of a company}
\usage{
quickfs_getmetric(tikr, metric, period = "FY-9:FY")
}
\arguments{
\item{tikr}{tiker of the company, char.}

\item{metric}{metric name, char.}

\item{period}{period ID (e.g., "FY-9:FY" or "FQ-19:FQ")}
}
\value{
financial data
}
\description{
Retrieve financial metrics of a company
}
\examples{

quickfs_getmetric("ILMN", "revenue", "FY-9:FY")

}
