% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickfs_getmetrics.R
\name{quickfs_getmetrics}
\alias{quickfs_getmetrics}
\title{Gets the metricks from quickfs}
\usage{
quickfs_getmetrics()
}
\description{
Gets the metricks from quickfs
}
\examples{

quickfs_getmetrics()

}
