% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickfs_getstatement.R
\name{quickfs_getstatement}
\alias{quickfs_getstatement}
\title{Retrieve annual financial statements}
\usage{
quickfs_getstatement(tikr, country)
}
\arguments{
\item{tikr}{tiker of the company, char.}

\item{country}{name of the country's headquarters}
}
\description{
Retrieve annual financial statements
}
