% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_time_price.R
\name{real_time_price}
\alias{real_time_price}
\title{Get real time stock price}
\usage{
real_time_price(tickers, api_key = api)
}
\arguments{
\item{tickers}{Required. The name of the tickers.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing real time stock price data.
}
\description{
Get real time stock price
}
\examples{
real_time_price(tickers = c("AAPL"),api_key = "42632c388365783ab0c341411f47ae80")
}
