% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_feeds.R
\name{rss_feeds}
\alias{rss_feeds}
\title{Get RSS feeds}
\usage{
rss_feeds(api_key = api)
}
\arguments{
\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing RSS feeds.
}
\description{
Get RSS feeds
}
\examples{
rss_feeds(api_key = api)
}
