% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cik_search.R
\name{search_cik}
\alias{search_cik}
\title{Cik search}
\usage{
search_cik(cik_name = NULL, api_key = api)
}
\arguments{
\item{cik_name}{Required. The cik name you want to search for}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results from your cik search.
}
\description{
Cik search
}
\examples{
search_cik(cik_name = "Berkshire", api_key = api)
}
