% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_performance.R
\name{sector_performance}
\alias{sector_performance}
\title{Get sector performance data}
\usage{
sector_performance(api_key = api)
}
\arguments{
\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing sector performance data.
}
\description{
Get sector performance data
}
\examples{
sector_performance(api_key = api)
}
