% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Install required non-CRAN packages from GitHub and elsewhere for the tutorials}
\usage{
setup(packages.cran = NULL, packages.github = NULL)
}
\arguments{
\item{packages.cran}{List of packages required from CRAN}

\item{packages.github}{List of packages required from GitHub}
}
\description{
\code{setup} installs all required non-CRAN packages from GitHub and elsewhere
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
