% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements_as_reported.R
\name{statements_as_reported}
\alias{statements_as_reported}
\title{Get statements as reported}
\usage{
statements_as_reported(ticker, type, period = "annualy", api_key = api)
}
\arguments{
\item{ticker}{Required. The name of the ticker.}

\item{type}{Rquired. Type of statements.
Valid values are:
"income"
"balance_sheet"
"cash_flow"
"full_financial"}

\item{period}{Required. The time period.
Valid values are:
"annualy" Default
"quarterly"}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing statements as reported.
}
\description{
Get statements as reported
}
\examples{
statements_as_reported(ticker = "AAPL", type = "income", period = "annualy", api_key = api)
}
