% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_news.R
\name{stock_news}
\alias{stock_news}
\title{Get stock news}
\usage{
stock_news(tickers, api_key = api)
}
\arguments{
\item{tickers}{Required. The name of the tickers.}

\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing stock news.
}
\description{
Get stock news
}
\examples{
stock_news(tickers = c("AAPL"), api_key = "42632c388365783ab0c341411f47ae80")
}
