% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorials.R
\name{tutorials}
\alias{tutorials}
\title{Return a data.frame with all currently available tutorials}
\usage{
tutorials()
}
\description{
\code{tutorials} returns a data.frame with all currently 
available tutorials.
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
