% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_scrape.R
\name{yahoo_scrape}
\alias{yahoo_scrape}
\title{Extracts information from YAHOO}
\usage{
yahoo_scrape(tikr, page = "", ntable = "all", stat = "all")
}
\arguments{
\item{tikr}{tiker of the company, char.}

\item{page}{page from which extracting information}
}
\description{
You can choose diferent \code{page} to scrape several types
of information. When nothing is defined, the summary page is scrapped by
default. There is also, \code{"key-statistics"}, which provides several
metrics, \code{"profile"}, which provides the names of key executives,
\code{"analysis"}, which offers different estimates for earnings,
revenues, etc.
}
