\name{financeR-package}
\alias{financeR-package}
\alias{financeR}

\docType{package}

\title{Financial Engineering with R}

\description{Financial Engineering with R - Support Package}

\details{
  \tabular{ll}{
    Package: \tab financeR\cr
    Type: \tab Package\cr
    Version: \tab 0.1-0\cr
    Date: \tab 2013-08-10\cr
    License: \tab GPL-3\cr
  }
}

\author{
  Ronald Hochreiter \email{ron@hochreiter.net}
  
  Maintainer: Ronald Hochreiter \email{ron@hochreiter.net}
}

\references{\url{http://www.finance-r.com/}}

\keyword{ package }

\examples{}
