% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tickers.R
\name{get_tickers}
\alias{get_tickers}
\title{Get tickers}
\usage{
get_tickers(api_key = api)
}
\arguments{
\item{api_key}{Required. Your API key.}
}
\value{
Returns a data frame of results containing all available tickers
}
\description{
Get tickers
}
\examples{
get_tickers(api_key = "42632c388365783ab0c341411f47ae80")
}
